chore(backend): use token-introspection client
