+++ packages/backend/src/app.ts
@@ -32,7 +32,6 @@ import {
   Grant,
   RequestAction
 } from './open_payments/auth/middleware'
-import { AuthService } from './open_payments/auth/service'
 import { RatesService } from './rates/service'
 import { SPSPRoutes } from './spsp/routes'
 import { IncomingPaymentRoutes } from './open_payments/payment/incoming/routes'
@@ -142,7 +141,6 @@ export interface AppServices {
   assetService: Promise<AssetService>
   accountingService: Promise<AccountingService>
   peerService: Promise<PeerService>
-  authService: Promise<AuthService>
   paymentPointerService: Promise<PaymentPointerService>
   spspRoutes: Promise<SPSPRoutes>
   incomingPaymentRoutes: Promise<IncomingPaymentRoutes>
