+++ packages/backend/src/open_payments/auth/middleware.test.ts
@@ -1,5 +1,6 @@
 import { generateKeyPairSync } from 'crypto'
 import { faker } from '@faker-js/faker'
+import { Client, ActiveTokenInfo } from 'token-introspection'
 import { v4 as uuid } from 'uuid'
 import {
   generateJwk,
@@ -11,7 +12,6 @@ import {
   createTokenIntrospectionMiddleware,
   httpsigMiddleware
 } from './middleware'
-import { Access, AuthService, TokenInfo } from './service'
 import { AccessType, AccessAction } from '../grant/model'
 import { Config } from '../../config/app'
 import { IocContract } from '@adonisjs/fold'
@@ -35,9 +35,9 @@ describe('Auth Middleware', (): void => {
   let deps: IocContract<AppServices>
   let appContainer: TestContainer
   let middleware: AppMiddleware
-  let authService: AuthService
   let ctx: PaymentPointerContext
-  const key = {
+  let tokenIntrospectionClient: Client
+  const key: ActiveTokenInfo['key'] = {
     jwk: generateTestKeys().publicKey,
     proof: 'httpsig'
   }
@@ -52,7 +52,7 @@ describe('Auth Middleware', (): void => {
       requestType: type,
       requestAction: action
     })
-    authService = await deps.use('authService')
+    tokenIntrospectionClient = await deps.use('tokenIntrospectionClient')
   })
 
   beforeEach(async (): Promise<void> => {
@@ -81,7 +81,7 @@ describe('Auth Middleware', (): void => {
   `(
     'returns 401 for $description access token',
     async ({ authorization }): Promise<void> => {
-      const introspectSpy = jest.spyOn(authService, 'introspect')
+      const introspectSpy = jest.spyOn(tokenIntrospectionClient, 'introspect')
       ctx.request.headers.authorization = authorization
       await expect(middleware(ctx, next)).resolves.toBeUndefined()
       expect(introspectSpy).not.toHaveBeenCalled()
@@ -96,7 +96,7 @@ describe('Auth Middleware', (): void => {
 
   test('returns 401 for unsuccessful token introspection', async (): Promise<void> => {
     const introspectSpy = jest
-      .spyOn(authService, 'introspect')
+      .spyOn(tokenIntrospectionClient, 'introspect')
       .mockResolvedValueOnce(undefined)
     await expect(middleware(ctx, next)).resolves.toBeUndefined()
     expect(introspectSpy).toHaveBeenCalledWith(token)
@@ -132,7 +132,9 @@ describe('Auth Middleware', (): void => {
         }
       })
 
-      const createTokenInfo = (access?: Access[]): TokenInfo => ({
+      const createTokenInfo = (
+        access?: ActiveTokenInfo['access']
+      ): ActiveTokenInfo => ({
         active: true,
         grant: uuid(),
         client_id: uuid(),
@@ -161,7 +163,7 @@ describe('Auth Middleware', (): void => {
             }
           ])
           const introspectSpy = jest
-            .spyOn(authService, 'introspect')
+            .spyOn(tokenIntrospectionClient, 'introspect')
             .mockResolvedValueOnce(tokenInfo)
           await expect(middleware(ctx, next)).rejects.toMatchObject({
             status: 403,
@@ -176,7 +178,7 @@ describe('Auth Middleware', (): void => {
         test('sets the context client info and calls next', async (): Promise<void> => {
           const tokenInfo = createTokenInfo()
           const introspectSpy = jest
-            .spyOn(authService, 'introspect')
+            .spyOn(tokenIntrospectionClient, 'introspect')
             .mockResolvedValueOnce(tokenInfo)
 
           await expect(middleware(ctx, next)).resolves.toBeUndefined()
@@ -205,7 +207,7 @@ describe('Auth Middleware', (): void => {
               }
             ])
             const introspectSpy = jest
-              .spyOn(authService, 'introspect')
+              .spyOn(tokenIntrospectionClient, 'introspect')
               .mockResolvedValueOnce(tokenInfo)
 
             await expect(middleware(ctx, next)).resolves.toBeUndefined()
@@ -229,7 +231,7 @@ describe('Auth Middleware', (): void => {
               }
             ])
             const introspectSpy = jest
-              .spyOn(authService, 'introspect')
+              .spyOn(tokenIntrospectionClient, 'introspect')
               .mockResolvedValueOnce(tokenInfo)
             await expect(middleware(ctx, next)).rejects.toMatchObject({
               status: 403,
@@ -283,7 +285,7 @@ describe('Auth Middleware', (): void => {
                 }
               ])
               const introspectSpy = jest
-                .spyOn(authService, 'introspect')
+                .spyOn(tokenIntrospectionClient, 'introspect')
                 .mockResolvedValueOnce(tokenInfo)
 
               await expect(middleware(ctx, next)).resolves.toBeUndefined()
@@ -306,7 +308,7 @@ describe('Auth Middleware', (): void => {
         test('returns 403 for super-action request', async (): Promise<void> => {
           const tokenInfo = createTokenInfo()
           const introspectSpy = jest
-            .spyOn(authService, 'introspect')
+            .spyOn(tokenIntrospectionClient, 'introspect')
             .mockResolvedValueOnce(tokenInfo)
           await expect(middleware(ctx, next)).rejects.toMatchObject({
             status: 403,
